[Presentation] Move DialogButtonInfo to Stride.Core.Presentation
